# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Drachenboot Manager API",
        "version": "2.5.0",
        "description": "API documentation for the Drachenboot Manager application with multi-team support."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/teams": {
            "get": {
                "summary": "Get All Teams",
                "description": "Fetches all teams ordered by name.",
                "tags": [
                    "Teams"
                ],
                "responses": {
                    "200": {
                        "description": "List of teams",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Team"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch teams",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Team",
                "description": "Creates a new team. Authenticated users only.",
                "tags": [
                    "Teams"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "Team name"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Name is required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create team",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Team": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique team identifier (CUID)"
                    },
                    "name": {
                        "type": "string",
                        "description": "Team name"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Team creation timestamp"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Last update timestamp"
                    },
                    "primaryColor": {
                        "type": "string",
                        "nullable": true
                    },
                    "showProRing": {
                        "type": "boolean"
                    },
                    "showProBadge": {
                        "type": "boolean"
                    },
                    "showWatermark": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message"
                    }
                },
                "required": [
                    "error"
                ]
            }
        },
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "authjs.session-token",
                "description": "Session-based authentication using NextAuth.js. Requires active login session."
            }
        }
    },
    "security": [
        {
            "sessionAuth": []
        }
    ],
...
}